GStreamer: fix QGstBus event loop integration
The QGstPipelinePrivate instance is owned by the GstPipeline and may be
unreffed/destroyed from a GStreamer thread. This results in the socket
notifier connection being disconnected from a worker thread, which is
not safe and will emit a warning.
In this case, we extend the lifetime of the QGstBus instance and marshal
the destructor through the application event loop.

Change-Id: I659a1896b3986ffab98f29eee9fe9edd2a17a5d8
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>
(cherry picked from commit 7c57d620d49878f90c7f7a67d2e1dc0f5bcca7fa)