Add code snippets for custom media input API
This change adds a cmake project containing snippets showing a
simplyfied way to setup and record custom media input with
QAudioBufferInput and QVideoFrameInput in pull mode.

Snippets from this project are added to the documentation of
QVideoFrameInput and QAudioBufferInput, together with a short
explanation.

The project is compilable and runnable to make it easy to test
and extend with more snippets in the future. The two main snippet
functions setupAndRecordVideo() and setupAndRecordAudio() are run
and exited automatically when starting the app. Resulting video and
audio files are saved to disk.

Task-number: QTBUG-126319
Change-Id: I2be62061d6d6fb8e4ec2b7490b50a663fd630bff
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>
(cherry picked from commit c01d223259a49f24b59eef8cbcceaf0a54909a0e)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>