Associate COM initialization and deinitialization with thread lifetimes
Use thread_local storage with a QComHelper instance to ensure that
CoUninitialize is called on the correct thread. All Qt Multimedia
objects requiring COM initialization will now request this from thread
local singleton, instead of managing COM initialization and
uninitialization within their own lifetimes. This approach ensures that
COM is uninitialized on the correct thread even if the objects are moved
between threads.

With this patch, CoUninitialize will be called when threads exits. This
means that for the main application thread, CoUninitialize will be
called during the destruction of static variables at process exit.

The Qt Multimedia documentation is improved, and explains how to avoid
common pitfalls related to COM initialization and uninitialization.

Fixes: QTBUG-129597
Pick-to: 6.8 6.5
Change-Id: I115f98f83a29f2aed7cf884a1017eb279fc5b7b7
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>