Create a new test library for multimedia test classes and functions
This patch introduces a new private static library MultimediaTestLib,
containing test utilities that are shared between tests. In this first
commit, we move helpers from tests/auto/shared into this library, and
move the implementation of qsinewavevalidator.h into
qsinewavevalidator.cpp. The latter is done to make sure we get a proper
C++ lib, not just a header-only library.

The library is put into qtmultimedia/src/multimediatestlib because this
follows the same pattern as for QuickTestUtilsPrivate, and ensures that
we do not get any build issues when building standalone tests.

We keep the licensing terms for the test utilities the same as our
tests, LicenseRef-Qt-Commercial OR GPL-3.0-only.

Task-number: QTBUG-129984
Pick-to: 6.8
Change-Id: Ica6400ba1d128c4695a2196a79d27ee18bb18056
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>