Move recorder tests to qmediarecorderbackend tests
Some tests that were previously added to qmediaframeinputsbackend tests
were really testing the QMediaRecorder. These tests are now moved to
the correct test suite.

There are still remaining tests involving the recorder in
qmediaframeinputsbackend tests, but these are the minimum required to
verify that the programmatic IO APIs work with the media capture
session. These tests should likely be moved to the mediacapturesession
tests once we have enabled them in CI.

Task-number: QTBUG-129984
Pick-to: 6.8
Change-Id: I1a01a92eea2c8da440e3d432c67637d0d5131851
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>