Android: Update for setting Audio Output device
Use new method (setCommunicationDevice[0]) to control audio output
device for Android API above 31.

Additionally use MODE_IN_COMMUNICATION when needed.
MODE_IN_COMMUNICATION is used to force the Android device to select a
lower-priority audio output. This mode is applied to the internal
speaker only when a Bluetooth or wired device is connected. For
Bluetooth, we check if a wired device is connected, and for wired
devices, we check if Bluetooth is connected.

[0]https://developer.android.com/reference/android/media/AudioManager#setCommunicationDevice(android.media.AudioDeviceInfo)

Task-number: QTBUG-128738
Task-number: QTBUG-126567
Pick-to: 6.8
Change-Id: I21beaf6d8fdca87260fda92bf761ef0bbac213cc
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>