Android: Increase BufferCount in AudioSink
BufferCount is a value used to initialize SLDataLocator_BufferQueue.
A low value increases the risk of buffer underflow. This happens when
the audio player consumes data faster than it can be replenished.

In this case, while playing sounds - the player may stop/start too
often. This can be observed with the "PlayerBase::stop() from IPlayer"
log that is constantly printed.

It seems that this issue may cause unexpected behavior - like force the
device to change audio output.

Fixes: QTBUG-128738
Fixes: QTBUG-126567
Pick-to: 6.8
Change-Id: Id3c0048459f1fc17d3118bba20a7742b64491b3e
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>