Document how to build FFmpeg on Windows with MSVC and MinGW-w64
Building FFmpeg on Windows is not well documented elsewhere, and to make
it easier to build custom builds of FFmpeg, the Qt documentation should
describe how to do this. In this patch we describe how to build FFmpeg
for 64 bits Windows on x64 platforms with Microsoft Visual Studio
compiler and with the MinGW-w64 toolchain.

Conflict in qtmultimedia-toc.qdoc was resolved by deleting the file.
This file does not exist on Qt 6.8.

Fixes: QTBUG-129489
Change-Id: I07da36ef06a4567d38be055ce67e57066db7afb4
Reviewed-by: Alexei Cazacov <alexei.cazacov@qt.io>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
(cherry picked from commit 4afdba0a97f9b3d172dd3ae2415c4cf5f789e3fa)