Only call find_package when building manual tests as separate project
find_package fails when building Qt because of a weakness in the CMake
infrastructure related to static plugins used in the mockbackend.

The solution is to only call find_package when we build manual tests as
indidual CMake projects, for example when opening the manual test in
Qt Creator.

This allows manual tests to be built together with auto-tests and the
Qt Multimedia source code by specifying QT_BUILD_MANUAL_TESTS=ON. This
helps when alternating between debugging through manual tests and
modifying the QtMultimedia code.

Pick-to: 6.8
Change-Id: I7ebbf618a9498833a641f394b937ef92f71e75cf
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>
Reviewed-by: Pavel Dubsky <pavel.dubsky@qt.io>
Reviewed-by: Nils Petter Skålerud <nils.petter.skalerud@qt.io>