Only reset audio client when format or buffer size has changed
This patch fixes an issue with QSoundEFfect on Windows, where short
sounds were cut. The sounds were cut because the IAudioRenderClient and
IAudioClient instances were released too early, before the sound
finished.

With this patch, the audio client and audio render client is kept alive
after being stopped. This allows it to drain completely and finish
playing the sound.

Fixes: QTBUG-130115
Fixes: QTBUG-112512
Pick-to: 6.8
Change-Id: I7a222a9c8c29ada4d11ff8b02e0be5f9ae102232
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Pavel Dubsky <pavel.dubsky@qt.io>