EVRCustomPresenter: guard the mixer with lock in ReleaseServicePointers
There are sporadic crashes due to nullptr dereference when accessing
the m_mixer in EVRCustomPresenter::processOutput(). This happens on
QMediaPlayer destruction because of a race condition. Thus, when
destroying the player, we call IMFMediaSession::Shutdown(), which
internally triggers EVRCustomPresenter::ReleaseServicePointers().
That in turn releases the m_mixer on player's thread. At the same time,
onSampleFree() callback might be called from WMF thread, which triggers
accessing the mixer in processOutput().

So we can simply fix this by adding a lock to the entire ReleaseServicePointer() call.

Fixes: QTBUG-128369
Pick-to: 6.8
Change-Id: Ifd294d8e4c1d5b2b90ead91a717005d9053964ef
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>