Turn tst_qmediaformat into a unit test
tst_qmediaformat tests functionality that benefits from unit testing
without actual backends. This will help us improving the test coverage
of the cross-platform parts of QMediaFormat, and can enable testing of
situations that are not reproducible with real backends.

This can enable testing of functionality that is only exposed for
example if the FFmpeg media backend is compiled without any video
codecs.

Task-number: QTBUG-130367
Pick-to: 6.8 6.5
Change-Id: I8ec985e1a91ecb499641066547318410b81fe25b
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Pavel Dubsky <pavel.dubsky@qt.io>