Rework tst_qmediaformat to improve test coverage
This patch reworks tst_qmediaformat to add test coverage of situations
that can not be tested with our current media backends.

The idea is that each test specifies which format and codecs are
supported, and that the tests verifies that the QMediaFormat behaves as
expected with those combinations. This allows us to test how
QMediaFormat behaves with backends that have "odd" format support.

Task-number: QTBUG-130367
Pick-to: 6.5
Change-Id: I55008c1ae30c12ebd6ae75cf2f592f737e52b7c9
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Pavel Dubsky <pavel.dubsky@qt.io>
(cherry picked from commit 833732c2bff72a881d21f162c830af0aad4a1511)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>