Clarify how to build Qt Multimedia without FFmpeg support
If FFMPEG_DIR is specified, configuration will fail even if
-no-feature-ffmpeg is specified, because we try to find FFmpeg before
the qt_feature(ffmpeg) is added in configure.cmake.

This patch updates the documentation to specify that disabling FFmpeg
requires omitting FFMPEG_DIR AND specifying -no-feature-ffmpeg. If
FFmpeg developer libraries are not present on the system, this is not
entirely true, because -no-feature-ffmpeg is then not required to
disable FFMpeg.

Fixes: QTBUG-130441
Pick-to: 6.5
Change-Id: I38eb722bb77445eebce919e5bd9abd42f0140df4
Reviewed-by: Pavel Dubsky <pavel.dubsky@qt.io>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
(cherry picked from commit 8c95842a8ec30dd13033acaf0a322388558b1817)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>