AVFMediaEncoder: call updateAudioInput() before starting recording
When the current device is disconnected, QAudioInput does not
react to this in any way, which leads to the fact that when
the device is reconnected and re-installed via calling
QAudioInput::setDevice(const QAudioDevice &device), the
QAudioInput::deviceChanged() signal will not be emitted,
which in turn will lead to AVFCameraService::audioInputChanged()
not being called and, as a consequence,
AVFCameraSession::updateAudioInput() will not be called, i.e.
new instances of AVCaptureDevice and AVCaptureDeviceInput will
not be created, which leads to the fact that the recording
will not be performed.

As a result of the fix, after disconnecting and connecting the
device (as in the case when the device was not disconnected),
an explicit call to AVFCameraSession::updateAudioInput()
occurs, which will lead to the fact that when the device is
reconnected, it will be recreated and recording will work
correctly.

Task-number: QTBUG-123073
Pick-to: 6.8 6.9
Change-Id: Icf3f94f272bfa00a5994d0eac075dbb67961ec42
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>