Android: Fix for default QCameraFormat
Before this commit, when default cameraFormat was not supported, the
first from the videoFormats list was returned as default.

The first item in the videoFormats can be the format with the highest
possible resolution. That may cause issues with the performance.
What is more - we could also wrongly assumed that default cameraFormat
was not supported in cases, when just min/max frame rate was different.

This commit returns the closest supported format as default (rather
than the first one in the list)

Fixes: QTBUG-129379
Pick-to: 6.5
Change-Id: I1ca3a422fa4489e64a0587b620fc159434764ee6
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Bartlomiej Moskal <bartlomiej.moskal@qt.io>
(cherry picked from commit 8383bc6a515e65fc3b7d0485a051bb9425ef2b8a)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>