GStreamer: Image capture - Simplify implementation
We used to use both buffer probes and a handover signal to obtain,
which are called in close succession, but with the same buffer. This
patch removes the handover signal and only relies on the buffer probes.

Resolution for 6.5:
* avoid QVideoFramePrivate::createFrame

Change-Id: I0cf6e3d961cab63079ebae76b9f1f5c6ae84802b
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>
(cherry picked from commit 7453f78c1800c6ea10efa3295cdb73742a608159)