Rename QFFmpeg::Codec to QFFmpeg::CodecContext
This class is wrapping an AVCodecContext, not an AVCodec, and does not
have any member functions that makes it a codec. CodecContext is
therefore a better name.

In this patch we also rename accessors from 'codec' to 'codecContext',
and rename variables from 'codec' to 'codecContext' for consistency.

Task-number: QTBUG-131270
Pick-to: 6.5
Change-Id: Ia762255fd3e6883ed01bda8f5cf8a42eebdc8864
Reviewed-by: Pavel Dubsky <pavel.dubsky@qt.io>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
(cherry picked from commit d477efb465230bdcdcf2128f4e9435f68eda703c)
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>
Reviewed-by: Lars Sutterud <lars.sutterud@qt.io>