Return optional from findBestAVValue when predicate is not fulfilled
The findBestAVValue family of functions returned a 'special value' when
the predicate could not be fulfilled. This required that calling code
knew what the 'special value' was to determine if a value was found.

We now return a std::optional<> when the value could not be found,
making failure explicit in the calling code. The calling code can choose
how to handle situations where no value was found, but does not have to
know if AV_SAMPLE_FMT_NONE, AV_PIX_FMT_NONE, or 0 has a special meaning.

This allow us to hide some complicated logic, including the generic
makeSpan function, and the InvalidAvValue inside the qffmpegcodec.cpp.
We still expose a AVPixelFormat specific version of the makeSpan
function for convenience, but we hide the generic implementation.

Task-number: QTBUG-131270
Change-Id: I935fb8117aff8a9f4b7cb2dfea4d090713a1c533
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
(cherry picked from commit b1ed833743b4cac94aeb0dd03cc0a43b217ae4fd)
Reviewed-by: Pavel Dubsky <pavel.dubsky@qt.io>
Reviewed-by: Nils Petter Skålerud <nils.petter.skalerud@qt.io>
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>