Use std::find_if to search in spans with predicate, and rename functions
Rename findAVValue to findIf, to match C++ terminology for searching a
range with a predicate. The function is reimplemented to use
std::find_if, which gives more familiar code than using findBestAVValue.

Also, rename hasAVValue to hasValue, because the function is now
operating on a general QSpan, and is not tied to how FFmpeg represents
arrays of values.

In the same go, the function arguments and template types are renamed
to better reflect what the input values are.

Task-number: QTBUG-131270
Change-Id: I0ea2286c56d4fca515b7c5514f11eccc63d74bd0
Reviewed-by: Pavel Dubsky <pavel.dubsky@qt.io>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
(cherry picked from commit 0368f04d39625c3f72fe6ef0092ba601ba070528)
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>