Make QFFmpegMediaFormatInfo easier to read
Shorten long lines, add some space between sections, and improve
documentation. Remove log statement that does not seem to be very
helpful. This requires removing the logging category because it is not
used and causes build error.

Manual but trivial conflict resolution on line 13 because static
logging category was not merged to 6.8. Logging category is no longer
present on dev, and the line was deleted on 6.8

Task-number: QTBUG-131270
Change-Id: I44ac204d387c3f65411ed32b3ce549da201d9141
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
(cherry picked from commit 188897a3f7ac97818001baefdfe68012c7032d0f)
Reviewed-by: Nils Petter Skålerud <nils.petter.skalerud@qt.io>