Android, Camera: Don't reset control members when restarting session
When processing of a frame takes too long, the Java side of the camera
code will trigger a full restart of the CameraCaptureSession.
Currently, this would also involve resetting control properties to
their default values. Ordinarily this should only happen when we
switch the camera-device. This issue commonly occurs during startup
or when under heavy CPU load, meaning control properties set early
in the program would be reset and go out of sync with the QCamera
object in the C++ code.

This patch makes it so that control-members in the Java-code is
explicitly only called during a QCameraDevice change.

Pick-to: 6.8 6.5
Change-Id: I4323c0a55cb4ee978e63d27a36892f23b52ca0ef
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
(cherry picked from commit c46a1a2da409aee797987826bb4ebb8664df2dcd)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>