Android, QtCamera2: Begin moving shared members into dedicated class
It is currently unclear how certain variables are synchronized across
the native C++ thread and the background-thread for handling
video frames and capture-progress. This makes it hard to determine
which members can be accessed from which thread at any gven time,
leading to harder maintainability.

The purpose of this patch is to improve structure and readability by
moving shared variabes into one class, and enforcing access with
the synchronize keyword.

There has been an attempt to keep the functional changes as minimal
as possible.

Pick-to: 6.8 6.5
Change-Id: I73f80f4dd3bdd8f62091efe80d2c07c3ecb85650
Reviewed-by: Bartlomiej Moskal <bartlomiej.moskal@qt.io>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
(cherry picked from commit 10684ddb2f6385dd4c88c297901c56da7b785837)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>