Add the endOfStreamPolicy property to VideoOutput
Note: This manual 6.8 cherry-pick varies significantly from the
original change. The original description is included below.
The cherry-pick backports the possibility the set the end of stream
behavior of VideoOutput, but instead of implementing it as a
QML property, the behavior is set through the environment variable
QT_VIDEOOUTPUT_EOS_POLICY. The documentation is also omitted,
making this as an udocumented feature in 6.8.

This backport is needed to fix QTBUG-129764. (check the comments)

The new env variable specifies the policy to apply when the video
stream ends. This occurs at the end of media playback or upon
deactivation of the camera, screen, or window capture.

- Clear the video output (default behavior):
QT_VIDEOOUTPUT_EOS_POLICY=ClearOutput

- Make the video output continue displaying the last frame:
QT_VIDEOOUTPUT_EOS_POLICY=KeepLastFrame

- In addition, VideoOutput.clearOutput() can be called to clear the
video output on demand.

-- Original change description: --

In Qt5 we had FlushMode for a simialar purpose.
The term Flush doesn't work well for this kind of functionality,
so it's better to call it according to what it's doing.

We want to have this API in qml only, because users have
an opportunity to filter frames through their own video sink in c++.

The method clearOutput() allows users to manually clear the last frame
if it has been kept and not needed anymore.

Task-number: QTBUG-110310
Change-Id: I238feb1592d5d055b2e04804e13a60b8c5307c78
Reviewed-by: Alexei Cazacov <alexei.cazacov@qt.io>
Reviewed-by: Nils Petter Skålerud <nils.petter.skalerud@qt.io>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
(cherry picked from commit a80fd4a0a6d38c6bb7876ee7df72ade0c9fcbd2a)