Only call avcodec_get_supported_config if codec type supports config
We log an error if we call avcodec_get_supported_config with a config
that does not match the codec type, for example, if we request the pixel
format from an audio codec. This spams the log during startup.

We fix this by returning empty from Codec functions that are not
relevant for the codec type.

In addition, we use the codec type to return early from
findAVPixelFormat if the codec is not a video codec, and the
dumpCodecInfo is adjusted to not log irrelevant info for the codec
type.

Task-number: QTBUG-131270
Change-Id: I524336b1fb8759b0472a3310991627765166d2d8
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Nils Petter Skålerud <nils.petter.skalerud@qt.io>
(cherry picked from commit 0da9a319073b1bfc531bf505efa26723f68f04e6)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>