Do not reset error state in QMediaPlayer::play()
The QMediaPlayer error state can be set when opening a file or url. This
is a common failure mode, and the error should not be reset until a new
source is set.

This patch leads to another potential issue, if the error occurs during
playback, and can be recovered by pressing play again. The
QMediaPlayer::error() will in this case continue to return error even if
playback succeeds. This recoverable failure mode is far less likely.

Fixes: QTBUG-115023
Change-Id: I383faadf2c210c3d47dc74bb9ee4bbbba2f3e6f4
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Lars Sutterud <lars.sutterud@qt.io>
(cherry picked from commit c5bf8a961906df50f8c475f52a8432a36dcbeacc)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>