Make QSoundEffect compilable without QT_network
This type of configuration is needed for the QtLite build.
The patch adds an opportunity to load samples by QFile.
The option is turned on when the QT_Network feature is not
available or in unit tests, where we want to check this
functionality.

No functional changes are expected when Qt_network is available.

Backporting to 6.8 and 6.9 to ensure codebase compatibility.

Pick-to: 6.8
Change-Id: I8c09eee053ace85f3fbd6547bdbd106bd4e333d5
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
(cherry picked from commit fbd172bba311c42eea5d149d893be3ab9613c2e8)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>