Test that QMediaFormat::resolveForEncoding() gives supported format
The QMediaFormat::resolveForEncoding() function attempts to reevaluate
the format/codec combination set by the application, to make sure that
the combination of container format and codecs are supported. This
patch verifies that the resulting combination is supported as determined
by QMediaFormat::isSupported.

To make the code more compact, we introduce a new allMediaFormats()
function that is used also in existing tests to make the code more
readable. We also rename the toString() function to enumToString()
because of name clash with QTestLib's toString() functions.

Task-number: QTBUG-131270
Pick-to: 6.8
Change-Id: I473a23b3efac584939afaa5b58a28586b13797f4
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
(cherry picked from commit 8392353a37c4e14028e8b0371c7f162013086d9a)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>