macOS, QAudioSource: Don't cache CoreAudio AudioDeviceID
Currently the QDarwinAudioSource class stores the CoreAudio
AudioDeviceID of the chosen device as a member variable. This value
changes whenever the pysical device is reconnected, but this class
never updates it's cached member variable accordingly. As such, it
fails to establish an audio-stream for a reconnected device.

This patch introduces a helper function that finds the most recent
AudioDeviceID for a given device unique-id. QDarwinAudioSource
will use this function whenever we try to establish an audio-stream.
This allows us to successfully recover a reconnected audio
input device on macOS.

Task-number: QTBUG-131929
Pick-to: 6.8 6.5
Change-Id: I1f2000187ac7481fe983b778b3b2079ffe9a3094
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
(cherry picked from commit 8bdefbb6412bdb6ccd9f56c8b6382a07174061f3)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>