GStreamer: QSharedHandle - fix move-to-self
When a shared handle is move-assigned, we cannot rely on QUniqueHandle
because it assumes that two instances point to different objects. This
may lead to memory leaks.

This also means that we cannot use QUniqueHandle for any refcounted
object and this patch replaces every use of QUniqueHandle with refcounted objects with QSharedHandle.

Fixes: QTBUG-132478
Pick-to: 6.8 6.5
Change-Id: I16035a32b9a458e731be88a1abb2d60af795c056
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>
(cherry picked from commit 0d9483007ecac12eb161cdf2442997bd979d1ffa)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>