QMediaMetaData: don't export parts of QHash
The newly-de-TP'ed asKeyValueRange() function exposes (in its return
type) QtPrivate classes uses to implement QHash, a class exporting of
which we consider to be an error.

So keep the function out of the ABI by (ab)using the
QT_POST_CXX17_API_IN_EXPORTED_CLASS macro.

This change is BiC on Windows, but the function was marked as TP in
6.8, so we can break BC here.

Found in API-review.

Pick-to: 6.8
Change-Id: Icd2d3789c73ade09f82ee38648f379a3adfcff7e
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>
(cherry picked from commit a0314d11a55b61e520c886d6f25eaa89404b91ed)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>