Remove Q_ASSERT from SamplePool::takeSample in EVRCustomPresenter
The assert verified that the sample pool was initialized when calling
SamplePool::takeSample(). It turns out that SamplePool::takeSample() can
still be called on an uninitialized SamplePool, causing the assert to
trigger.

We suspect that the reason is that the presentation clock is not stopped
when releasing the EVR presenter, or that already queued timer events
from a previous presentation clock may be forwarded to the new EVR
presenter through EVRCustomPresenter::OnClockStart and
EVRCustomPresenter::OnClockRestart. We were unfortunately not able to
find a way to stop the clock and drain the Media Foundation's event
queue from such events in a synchronous manner.

Returning S_FALSE from EVRCustomPresenter::processOutput() when the
SamplePool is empty still seems like the right thing to do, as this will
exit the processOutputLoop() function and return control to the media
foundation. Note that this makes the EVRCustomPresenter deviate from the
Microsoft Windows SDK sample code that has this assert.

Amends 21df2fa0a6f493df6ffb3b3efe3424e21500c5ea.

Fixes: QTBUG-130299
Pick-to: 6.8
Change-Id: Ic48b0afcfd23f11687536055b2707c764f1c152f
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Pavel Dubsky <pavel.dubsky@qt.io>
(cherry picked from commit 693aa2c4fb283120485a183b5eba6231a690ebac)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>