Android, Camera: Move focus-mode settings into dedicated function
Currently we are setting the camera focus mode manually in multiple
places in the code. This can be hard to maintain long-term and makes
the code less readable. This is especially problematic once
we introduce support for more focus-modes on Android.

This patch moves the logic of the focus-mode into a dedicated
function so we can always set it correctly.

Task-number: QTBUG-113155
Pick-to: 6.8
Change-Id: I5da4a2fd28dd254c5ee1f3955d95a2facc210687
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Bartlomiej Moskal <bartlomiej.moskal@qt.io>
(cherry picked from commit 433396857dba1b9b14888d328ac7703b24a84aef)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>