QImageCapture: Don't fire fileFormatChanged when property was unchanged
Currently we fire the fileFormatChanged signal regardless of whether
the property was actually modified. This can happen such as when the
user assigns a FileFormat that is not supported.

This patch makes setFileFormat only fire the fileFormatChanged signal
if the property actually changed.

Fixes: QTBUG-133250
Pick-to: 6.8 6.5
Change-Id: If3beb2508060d58686d63618c17f9cc367f804df
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
(cherry picked from commit f089c5213dbccb82eb9652528caf6bfa94a213d9)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>