Propagate "redOrAlphaIndex" to shaders
The patch prepares the code for reusing the shaders logic for both,
normal shaders, and fallback shaders (*_a.frag).

The idea is having the index of the actual component index in
shaders, instead of enquiring it by the component name, .a or .r.

Pick-to: 6.8
Change-Id: I719a2a99318c99d28e6fe938858e5c31e35820fd
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>
Reviewed-by: Mikko Hallamaa <mikko.hallamaa@qt.io>
(cherry picked from commit 1370e3204ea1ccbee7ffd43581de835fbea84247)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>