Handle QGuiApplication::applicationStateChanged in proper threads
We should handle the signal applicationStateChanged in the rhi's
thread, because it deletes the rhi and the fallback surface,
otherwise, it leads to unexpected behavior.

The change touches the Android platform, as the signal
is handled only there.

Pick-to: 6.8
Change-Id: I8f370ba1efc344c8d48d7291f586d2780a866714
Reviewed-by: Bartlomiej Moskal <bartlomiej.moskal@qt.io>
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>
(cherry picked from commit 5876e63a5424268d6f15ce5e96e513d1aa8b07cd)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>