CoreAudio: listen to device disconnect
When a device is disconnected, the coreaudio callback is not served
anymore, but we also don't get an error message. To do so, we need to
subscribe to the kAudioDevicePropertyDeviceIsAlive property. This allows
us to track device disconnects so that we can raise an IOError.

Pick-to: 6.8
Change-Id: I67b172a10b56cd6cb82af7f3796a690625960efc
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>
(cherry picked from commit eb0c42238240cc8ef3537bed26a5977edd844873)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>