Fix freeze in GStreamerMediaPlayer when destroying QMediaPlayer
In QGstreamerMediaPlayer::setAudioOutput, we call
QGstElement::finishStateChange to let the QGstPipeline finish any
pending state transition. This is a workaround to avoid assertion
failures in GStreamer during playback. However if it is in stopped
or paused state, it can cause the application to freeze for a preset
timeout of 5 seconds.

If we are about to destroy the pipeline, we can skip waiting for the
state transition. This change therefore adds a flag to
QPlatformMediaPlayer which is set to true in ~QMediaPlayer and used
as a !condition to call finishStateChange.

This is a temporary workaround to fix freezing when destroying
QMediaPlayer while paused or stopped. A future solution should
handle other scenarios as well.

Fixes: QTBUG-133135
Pick-to: 6.8
Change-Id: Id8cbc4f20da7f4edecc3f365273191faa19d19ff
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
(cherry picked from commit 79e016539613150f92555ff42966e02b3d69e2dd)