MediaCodec: Do not set level value for H265 Video Codec
The level value in FFmpeg is used to specify the compression constraints
for an H.265/HEVC-encoded video. While values mapped in the array seems
to be correct[0], they are not recognized by avcodec_open2() method.

It seems that this is a regression[1] on ffmpeg side. It is already
fixed[2], but the ffmpeg package from 7.1 tag doesn't contain the fix.
The fix is already on release/7.1 branch.

This fix uses avcodec_version() method to filter out not working FFmpeg
versions.

[0]https://github.com/FFmpeg/FFmpeg/blob/2066c5526d27168144db059c344df58ed5942aa3/libavcodec/mediacodecenc.c#L1241
[1]https://github.com/FFmpeg/FFmpeg/commit/7753a9d62725d5bd8313e2d249acbe1c8af79ab1
[2]https://github.com/FFmpeg/FFmpeg/commit/020d9f2b4886aa620252da4db7a4936378d6eb3a

Pick-to: 6.9 6.8
Fixes: QTBUG-133773
Change-Id: I4f5051c14bf0523e383a7a0969cc50b131a96b2b
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>