Add a configure option for native MediaCodec for Android
We have deprecated the MediaCodec for Android, and adding it as a
"native_android_backend" feature can allow us to not build it. In
addition, the configure summary now shows that the backend is
deprecated.

After this patch we still build the Android backend by default. To
disable building the native Android backend, configure Qt with configure
flag -no-feature-native_android_backend or CMake option
-DFEATURE_native_android_backend=OFF

Task-number: QTBUG-125370
Pick-to: 6.9 6.8
Change-Id: I5172317439b3ff1dd4ac22e7b1fd5b2e7dc6c454
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>