Document the native Windows media backend as deprecated as of Qt 6.10
With Qt 6.10, we no longer recommend using the native Windows media
backend built on top of Windows Media Foundation (WMF). Instead,
applications should use the FFmpeg media backend.

This patch updates the Qt landing page to mark the native Windows
backend as deprecated. Configure scripts are updated to make it possible
to opt-out of building the native media backend. The Windows media
backend is still built by default, and is still part of the Qt online
installer.

Since the FEATURE_wmf feature flag is also required for building the
WMF audio backend and helper classes, we introduce a new feature flag
to control the Windows media backend specifically. Note that the built-
in WMF audio backend is not deprecated.

[ChangeLog][Document the native Windows media backend as deprecated as
of Qt 6.10]

Fixes: QTBUG-134033
Change-Id: I7e390f4b07f32fde7d9287ef8d2583dfb53f5eac
Reviewed-by: Pavel Dubsky <pavel.dubsky@qt.io>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>