Remove QRhi* as argument to private rhiChanged() signal
The QRhi instance passed to the QVideoSink and implementation classes
is set in the render thread, and requires holding a mutex to be able to
use it. Removing the QRhi instance from the signal makes it harder to
introduce concurrency related bugs.

Task-number: QTBUG-133033
Pick-to: 6.8
Change-Id: I07593146635b3d032713315e4b37279c6c58b843
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
(cherry picked from commit 3a8516ecdbb3d82fd6c758657df97556f64809e7)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>