tst_qvideoframecolormanagement: Fix testing for multiple fallbacks
Testing rhi-based video frame conversion with various combinations of
excluded rhi textures currently causes exponential growth of rendering
modes to be tested and compared against when skipping specific
exclusions.

This patch refactors the test to allow testing of any combination of
excluded texture formats, as long as their exclusion creates a unique
fallback chain for the tested pixel format on the tested platform.
Excluded rhi texture formats are now flags that allow bitwise operations
for checking a specific exclusion in a tested combination of exclusions.

Pick-to: 6.10 6.9 6.8
Change-Id: Id97ecf3210196033d6f84761337c822608001921
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>