Audio: rework default device fallback
CoreAudio used to fall back on the default audio input/output in case of
a null device. This is inconsistent with all other platforms and dead
code, since null devices are already filtered out in
QPlatformAudioDevices.
Furthermore, the implementation used the "first" device as fallback, not
the "default" device.

This patch removes the coreaudio specific code and updates the fallback
code to use the "default" device first before and only use the "first"
device as last resort

Pick-to: 6.8 6.9
Change-Id: If04bd0c287fd356c31c6e736b5ce58dea0663b85
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>