AudioRecorder Example: Fix crash on permissions denied
The AudioRecorder example could crash because QMediaRecorder is only
created when microphone permissions are granted. However, the code
assumes that m_audioRecorder is always set, potentially leading to a
null pointer dereference.

This commit adds checks for m_audioRecorder and displays a message
in the status bar when permissions are missing.

Pick-to: 6.9 6.8 6.5
Fixes: QTBUG-134046
Change-Id: I184f2b795825d2888b504ad03c0c9384e588518b
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>