QML, Camera: Expose RESET functions for device and format
In QML, there is currently no way for the user to go back to the
default format if they ever chose another one.

This patch introduces RESET bindings to these QML properties so that
the user can assign 'undefined' in order to request the QCamera object
to select the default value, based on what is available.

[ChangeLog][Multimedia][Camera] QML property cameraDevice can now be
reset in order to select the default device.
[ChangeLog][Multimedia][Camera] QML property cameraFormat can now be
reset in order to select the default format.

Task-number: QTBUG-134445
Change-Id: If8a173c50718b94329d1e2585754070cacb4bf05
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>