Tests: rename tst_qml to tst_multimedia_qml
We already have a tst_qml in Qt Declarative, and multiple targets with
the same name break toplevel builds.

This has so far not been noticed because this subdirectory is not added
by the parent project. Let's not wait for it.

Change-Id: I4a53a5d2b3c479f4638ee57208548f19dcf6773d
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Pavel Dubsky <pavel.dubsky@qt.io>
(cherry picked from commit ed16142e451fecf08b53096017dce73faef98a47)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>