iOS, FFmpeg: Modify FindFFmpeg.cmake to look for .xcframework files
Currently the FindFFmpeg.cmake script looks for the presence only
of .dylib files in order to determine the validity of the FFmpeg
installation. On iOS, we no longer deploy these file-types directly for
shared libraries. We only deploy .xcframework files, which in turn
contain the .dylibs. This creates a situation where FindFFmpeg can
succeed even when .xcframework files are missing from the FFmpeg
installation.

This patch allows FindFFmpeg.cmake to look for .xcframework files when
on iOS. This will also allow us to stop installing raw .dylibs as part
of FFmpeg builds.

Fixes: QTBUG-135172
Pick-to: 6.9 6.8
Change-Id: I215700dea7e2f8ee7e513c8dbaa0c850d4f41a03
Reviewed-by: Timur Pocheptsov <timur.pocheptsov@qt.io>