Camera, Darwin: Don't expose suspended camera-devices
Currently we expose camera-devices that are marked as suspended. This
means they are connected but otherwise not possible to use, i.e
currently being used exclusively by another app. Another example is
the Macbook camera when the lid is closed.

This patch makes us no longer expose these devices as usable, and track
their suspended state so we can expose them when they are no longer
suspended.

Fixes: QTBUG-132167
Pick-to: 6.9 6.8
Change-Id: Ic3b7c678591d89b25537886a52a8013e39bdf15f
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>