Fix integer underflow regression for mp3 files with attached_pic
Calculating (avPacket.pts - avStream.start_time) causes the integer to
wrap around to a very large number if avPacket.pts == AV_NOPTS_VALUE,
which is the case for the first packet of mp3 files with an
attached_pic (thumbnail image). However that packet comes from the
thumbnail image stream, which isn't relevant for playback.

This change makes sure isPacketWithinStreamDuration is only called
while demuxing streams that are relevant, i.e. present in the
m_streams map. In addition it makes sure the same function returns
true if it is called with a AV_NOPTS_VALUE packet. Since that's
unexpected and could lead to undefined behavior, a qWarning is
logged.

Also cleans up some code duplication in Demuxer::doNextStep, and
adds a test that checks if such files are being played back.

Fixes: QTBUG-135307
Pick-to: 6.9 6.8
Change-Id: Id5cedcd7a2b6097340262f9f9bb43df73035d9e0
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>