Darwin, camera: Simplify device switching logic
The current code requires us to initialize the m_videoDataOutput
object before we run setPixelFormat. This object is initialized
deferredly, during setCameraDevice, but will fail if permissions
have not been granted. In the context of QML, permissions will
commonly be granted after QCamera initialization and setup.

This patch fixes an issue where we are unable to apply format
if camera permission was granted after the setup of the QCamera
object. An additional side-effect is that it simplifies the
logic of device-switching by moving static components into
the constructor.

Task-number: QTBUG-134430
Pick-to: 6.9 6.8
Change-Id: I2b5af71612850c8e121c373b87884cade0b8020e
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>