Darwin, Camera: Configure AVCaptureSession with device during setActive
There is currently a bug where, if permissions are granted after
setting up the initial QCamera object, the camera will fail to work
correctly once permissions have been granted and the camera is
requested to go active. This is because connecting the camera-device
to the AVCaptureSession only happens during setCameraDevice, which
in this case is unable to configure the device if permissions are not
granted.

This patch fixes this issue by running the configuration step during
setActive.

Fixes: QTBUG-135256
Pick-to: 6.8
Change-Id: I5967c3cdc2a1e86cb6bc26ed3a75f84a4910bd8e
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
(cherry picked from commit 42fea923655fb26e6c60ba21c51062afe9f0bfe7)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>