Darwin, camera: Remove error-prone AVCaptureDevice getter
The Darwin FFmpeg camera backend relies on whether a
AVCaptureDeviceInput is configured for the on-going AVCaptureSession to
determine how to grab the current AVCaptureDevice. This is error-prone
because this might not always be the case. We cannot instantiate the
AVCaptureDeviceInput whenever permissions are not granted, as that will
implicity trigger a permission request to the end-user, which is
undesirable. This means some of our configuration steps will
misinterpret the AVCaptureDevice as disconnected when permissions are
not granted but we are still configuring our QCamera.

This patch removes the AVCaptureDevice-getter in the FFmpeg backend.
This allows code to fallback to the getter in the QAVFCameraBase
superclass which has the same name and the more robust implementation.

Task-number: QTBUG-135256
Change-Id: I1a38dafb71309089e383db5e2fab830e2da91f80
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
(cherry picked from commit 6a52a7c1f5046a7f615fbc97b106fffba7ba29be)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit 596d097d8d701fe82accfe950c70d5e3fe54c034)