Darwin, Camera: Solve memory leak related to rotation tracking
The current uses a method 'updateRotationTracking()' to setup state
related to tracking the rotation of the AVCaptureDevice. This method
either refreshes or clears the state depending on whether the camera is
active or not. During destruction of the QAVFCamera object, the camera
is still in active state, and therefore sets up new rotating-tracking
objects during destruction and never cleans it up.

This patch solves this memory leak by introducing a method
'clearRotationTracking()' and calls it during destruction.

Pick-to: 6.10 6.9 6.8
Change-Id: I1f457ae62aa781dff48b50fb1f744e71da6a7972
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>