QVideoFrame: Fix calculation of UV line stride
UV line stride in planar YUV and YV formats is calculated from overall
frame size in bytes, frame height, and the known Y-plane stride.

The overall size comes from bytesRequired in texture description, which
aligned the U and V plane heights to the nearest even number. This
caused an inaccuracy in the calculation for UV stride for resolutions
where the U and V planes have odd heights.

The increase of horizontal scaling in packed texture formats (fallback
for YUV420p10) increased this error, and caused an extra pixel to be
added to the stride.

This patch replaces the alignment in bytesRequired of these formats, to
align the sum of U and V plane heights instead, and re-enables
previously failing tests.

Pick-to: 6.10 6.8
Change-Id: I9c8912eeba4729b0117bdd7e87a76c43569cb87a
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>
(cherry picked from commit ec793ce3541dfdec0efa5a4f55d64c817a245e5d)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>