Android: Mark QtScreenCaptureService as public
This ammend commit 2f19560522d4044c626324eebd2515b2e418592d.

This patch introduced a regression that caused a crash whenever this
Service was being instantiated by the OS. This Service class needs to
be accessible and visible to the system class
android.app.AppComponentFactory. This is because only the Android
system is allowed to instantiate Service classes.

Fixes: QTBUG-136191
Pick-to: 6.9 6.8
Change-Id: I77889078dd0c4dbad1ac986d23821c671ca8dd9d
Reviewed-by: Bartlomiej Moskal <bartlomiej.moskal@qt.io>